[FLINK-17926] Singleton the command used to build website using Docker
Before we can build the website using `./docker/run.sh` and the command `docker run --rm --volume="$PWD:/srv/flink-web" --expose=4000 -p 4000:4000 -it ruby:2.5 bash -c 'cd /srv/flink-web && ./build.sh -p'` (which in the the page `Contribute to the Website`).

Now we remove the `./docker` directory, and using the command as the only official way to build Flink website.

Also add the command to README.md, so that we can easily find it.

This closes #342